+++ Tests/SwiftLintFrameworkTests/RulesTests.swift
@@ -220,43 +220,6 @@ class RulesTests: XCTestCase {
         verifyRule(TodoRule.description, commentDoesntViolate: false)
     }
 
-    func testTrailingComma() {
-        // verify with mandatory_comma = false (default value)
-        verifyRule(TrailingCommaRule.description)
-
-        // verify with mandatory_comma = true
-        let mandatoryCommaDescription = RuleDescription(
-            identifier: "trailing_comma",
-            name: "Trailing Comma",
-            description: "Trailing commas in arrays and dictionaries should be enforced.",
-            nonTriggeringExamples: [
-                "let foo = []\n",
-                "let foo = [:]\n",
-                "let foo = [1, 2, 3,]\n",
-                "let foo = [1, 2, 3, ]\n",
-                "let foo = [1, 2, 3   ,]\n",
-                "let foo = [1: 2, 2: 3, ]\n",
-                "struct Bar {\n let foo = [1: 2, 2: 3,]\n}\n",
-                "let foo = [Void]()\n",
-                "let foo = [(Void, Void)]()\n",
-                "let foo = [1, 2, 3]\n",
-                "let foo = [1: 2, 2: 3]\n",
-                "let foo = [1: 2, 2: 3   ]\n",
-                "struct Bar {\n let foo = [1: 2, 2: 3]\n}\n",
-                "let foo = [1, 2, 3] + [4, 5, 6]\n"
-            ],
-            triggeringExamples: [
-                "let foo = [1, 2,\n 3↓]\n",
-                "let foo = [1: 2,\n 2: 3↓]\n",
-                "let foo = [1: 2,\n 2: 3↓   ]\n",
-                "struct Bar {\n let foo = [1: 2,\n 2: 3↓]\n}\n",
-                "let foo = [1, 2,\n 3↓] + [4,\n 5, 6↓]\n"
-            ]
-        )
-
-        verifyRule(mandatoryCommaDescription, ruleConfiguration: ["mandatory_comma": true])
-    }
-
     func testTrailingNewline() {
         verifyRule(TrailingNewlineRule.description, commentDoesntViolate: false,
                    stringDoesntViolate: false)
@@ -396,7 +359,6 @@ extension RulesTests {
             ("testSwitchCaseOnNewline", testSwitchCaseOnNewline),
             ("testSyntacticSugar", testSyntacticSugar),
             ("testTodo", testTodo),
-            ("testTrailingComma", testTrailingComma),
             ("testTrailingNewline", testTrailingNewline),
             ("testTrailingSemicolon", testTrailingSemicolon),
             ("testTrailingWhitespace", testTrailingWhitespace),